DCPP Solver
Other than Fleury's and one final architecture change, should be done
with first pass DCPP (pre-performance optimization).

+++ src/oarlib/core/Arc.java
@@ -17,7 +17,7 @@ public class Arc extends Link<DirectedVertex>{
 	 * @param endpoints
 	 * @param cost
 	 */
-	public Arc(String label, Pair<DirectedVertex> endpoints, double cost) {
+	public Arc(String label, Pair<DirectedVertex> endpoints, int cost) {
 		super(label, endpoints, cost);
 	}
 	/**
