+++ src/oarlib/core/Edge.java
@@ -9,7 +9,7 @@ import oarlib.vertex.impl.UndirectedVertex;
  */
 public class Edge extends Link<UndirectedVertex>{
 
-	public Edge(String label, Pair<UndirectedVertex> endpoints, double cost) {
+	public Edge(String label, Pair<UndirectedVertex> endpoints, int cost) {
 		super(label, endpoints, cost);
 	}
